+++ src/syssense/__init__.py
@@ -1,4 +1,4 @@
 """SysSense - Monitor inteligente do sistema para Fedora/GNOME."""
 
-__version__ = "0.2.1"
+__version__ = "0.3.4"
 __author__ = "Marcos"